refactor: move registry alias handling to dockerfile manager
